added a fragment to display side scrolling recipe information retrieved from the web service in MainActivity  - currently displays newest created, trending, and saved recipes
added fragment to display a snippet of the Recipe after selecting it from the side scroll in MainActivity

added HttpRecipe to deal with REST calls for recipe information

fixed recipe category to show up properly
updates the stored recipe category list in RecipeListPresenter on category added or removed

removed SectionEntity references from Room database

added foreign keys to all bridge tables with cascading delete

changes all references of measurement types to a long id
 - added UnitOfMeasurementEntity to database to reference  the measurement type through a foreign key instead of a String

removed CountTuple class

changed references of price and quantity to float
 - Changed to Float inside database and ability to display the floats

converted Date to TimeStamp in recipe to be more easily compatible with mssql DateTime2

added description to Recipe that is passed to and from RecipeEntity

added OnlineRecipe to deal with Recipes retrieved from web service

changed the look of the recipe tags background to be more square, and close together inside the RecyclerView

+++ app/src/main/java/com/habbybolan/groceryplanner/database/entities/IngredientEntity.java
@@ -15,7 +15,7 @@ public class IngredientEntity {
     private long ingredientId;
     @NonNull
     private String ingredientName;
-    private int price;
+    private float price;
     @ColumnInfo(name="price_per")
     private int pricePer;
     @ColumnInfo(name="price_type")
@@ -23,7 +23,7 @@ public class IngredientEntity {
     @ColumnInfo(name = "food_type")
     private String foodType;
 
-    public IngredientEntity(long ingredientId, @NonNull String ingredientName, int price, int pricePer, String priceType, String foodType) {
+    public IngredientEntity(long ingredientId, @NonNull String ingredientName, float price, int pricePer, String priceType, String foodType) {
         this.ingredientId = ingredientId;
         this.ingredientName = ingredientName;
         this.price = price;
@@ -44,7 +44,7 @@ public class IngredientEntity {
     public String getIngredientName() {
         return ingredientName;
     }
-    public int getPrice() {
+    public float getPrice() {
         return price;
     }
     public int getPricePer() {
